pyproject.toml
[project]
name = "resume-gap-analyzer-mcp"
version = "1.0.0"
description = "MCP server that identifies gaps in resumes compared to job descriptions using OpenAI"
requires-python = ">=3.9"
dependencies = [
    "fastmcp>=0.1.0",
    "openai>=1.0.0",
    "python-docx>=0.8.11",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.ruff]
line-length = 100
target-version = "py39"
